Added generic class SuperArray: simple generic array class we can use across the entire project
Added generic LinkedList and Node classes: This is mainly required for
the adjacency list but as it is generic can be re-used across the
project to store any object type.

Added Vertex class: this holds the vertex information (eg. the SIRState
and an index value to an array)

Updated AbstractGraph: Added a hashmap to be used across all graphs

Updated AdjacencyList: adding and deletion methods implemented